InsertId = undefined in iOS
Problem occurs when you insert values into different tables, but the
insertId is the same.  For example, if the last record you inserted was
into tableA and the rowId = 1, then you insert a record into tableB
which also returns a rowId = 1, hasInsertId remains NO and returns
undefined. Also reported as issue #63