Automate the unit tests and improve test coverage
This adds some structure to make it easier to
automate the unit tests.  With this PR, you can simply
run ./bin/test.sh ios or ./bin/test.sh android
and it'll automatically spin up an emulator and
run the QUnit tests.

I also went ahead and refactored the tests so that
you can now test both the native `openDatabase` and
this plugin.  That'll come in handy when you want to
test that the plugin behavior matches the native
behavior.