Propagate statement failures to transaction failures.
According to the W3C spec, if a transaction failure callback *does not return false* then the error is propagated to a transaction failure. This means that if a transaction returns true, but more importantly, if it has no return value (undefined) then the error must propagate to the transaction. This fix explicitly requires false to correct the propagation.

A subsequent fix will address the fact that the original statement error is not propagated and is replaced here, but that is unrelated to this issue.