Fix truncation in iOS query result string encoding.
This fix enables end-to-end encoding of binary data using strings.

Strings with Unicode values like `\u0000` were being truncated because the string constructor was using a null terminator to determine size instead of the stored byte length. The fix uses a different constructor and the stored byte length so that the full length string is returned.