cordova-sqlite-ext doc updates
NOTE: cordova-sqlite-ext 0.10.6 has incorrect -DSQLITE_THREADSAFE=1
build flag on iOS due to a mistake in the merge. This will be changed
to SQLITE_THREADSAFE=2 for iOS in the next release.

ADDITIONAL NOTE: plugin.xml has duplicate macOS ("osx") platform
section, also due to a mistake in the merge. The duplicate macOS
platform section will be removed from the next release.

ref: litehelpers/Cordova-sqlite-storage#703